Stop adding bluebook(1) in all sites
Since we've deployed a GitHub Pages site, and link to that from the
footer now, we don't need to include this every time we're used. We'll
build it as a separate step in the Pages build and it'll get pulled in
like any other page to ensure consistency without any special-handling.